requirements.txt
flask
discord.py
requests
python-dotenv
aiohttp